Implement lazy-loading for the ReferenceObject cache
This feature is intened to be used when running small tests that do not
require all cacheable objects to be in the cache. Loading everything in
to the cache can be time consuming, so lazy loading ensures that you
will only need to wait for objects that your test will use.

You *could* use this in production (or for a large set of tests that
cover all cacheable objects) but there would be little point to that.