Get some work done on the ENSEMBL file processor.
Introduced the concept of a "globbed" file processor because I think it
might be easier if ENSEMBL file processors can operate on file globs
instead of explicit names.

Still a bit of work to do to finish this up, hopefully will get it done
on Monday.

Also, the UniProt file processor can probably be refactored similar to
how the Ensembl File processor works.