Once again, working with ENSEMBL complicates things...
To be able to generate cross-references properly for ENSEMBL, we need to
keep track of the original ENSP with all the intermediate values along
the way. 
Also, introduced a new interface for all classes that do bulk reference
creation so there can be a common interface type instead of coding to
specific classes.