Duplicate identifiers report: check referres in code
It is possible to construct an SQL query that will consider referrers
when checking for duplicates, BUT it ran very slow when I tried to run
it on reactomerelease so I think it might be best if I run the referrers
check in Java, on a per-identifier basis rather than run the check as a
part of the query.

Also:
 - updated unit test
 - Update testing depenedencies in pom file; added some metadata to pom
file.