Allow the ReferenceCreator for OpenTargets to search geneName
The file from OpenTargets has GENE NAMES mapped to ENSEMBL identifiers,
so SimpleReferenceCreator needed to be modified so that the source
identifying attribute could be more dynamic, but it still defaults to
ReactomeJavaConstants.identifier.

Also: Allow SimpleReferenceCreator to build a list of "things to create" using parallel streams on the input sourceReferences, and then use that list later to populate the database. Offers a slight speed-up.